# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto controle de Filmes - SENAI
  description: |-
   API destinada ao controle de Filmes para o projeto "Bem Legal" do SENAI
  termsOfService: http://swagger.io/terms/
  contact:
    email: seunome@email.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.4.26
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: Endpoint para enviar os dados do filme a serem cadastrados
      requestBody:
        description: JSON contendo os dados do novo Filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Successful operation
        '400':
          description: Invalid input
        '500':
          description: Error Internal Server (Controller / Model)
        '415':
          description: Error Content Type
    get:
      tags:
        - Filme
      summary: Lista todos os Filmes
      description: Listagem de todos os Filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Aluno'          
           
        '404':
          description: Not Found
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Buscar um Filme pelo ID
      description: Retorna um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Invalid ID supplied
        '404':
          description: not found
        '505':
          description: Error Internal Server (Controller /  Model) 
    put:
      tags:
        - Filme
      summary: Atualizar um filme pelo ID
      description: Buscar e atualizar um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um Filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Aluno'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
        '415':
          description: Invalid Content-Type
        '500':
          description: Error Internel Server (Controller / Model)
       
    delete:
      tags:
        - Filme
      summary: Delete um filme pelo ID
      description: Buscar e Excluir um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para buscar o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Not not found
        '500':
          description: 'Error Internal Server (Controller / Model)'
  /v1/senai/locadora/genero/:
    post:
      tags:
        - Gênero
      summary: Adiciona um novo Gênero
      description: Endpoint para enviar os dados de Gênero a serem cadastrados
      requestBody: 
        description: JSON contendo os dados de novo Gênero a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/componente/schemas/Genero'
        required: true
      responses:
        '201': 
          description: Successful operation
    get:
      tags:
        - Gênero
      summary: Buscar um Gênero pelo ID
      description: Retorna um Gênero pelo ID
components:
  schemas:
    Filme:
      required:
        - nome
        - sinopse
        - data_lancamento
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sinopse:
          type: string
        data_lancamento:
          type: string
        duracao:
          type: string
        capa:
          type: string
        valor:
          type: integer
        avaliacao:
          type: integer
  requestBodies:
    Filme:
      description: Objeto que deve ser enviado no Body das requisicoes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Filme'
    